requirements.txt
Django==1.10.6
appdirs==1.4.3
defusedxml==0.5.0
django-allauth==0.31.0
django-autoslug==1.9.3
django-humanize==0.1.2
humanize==0.5.1
oauthlib==2.0.1
packaging==16.8
pyparsing==2.2.0
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
